requirements.txt
python-dotenv==0.20.0
ccxt==1.91.67
pandas_ta==0.3.14b0
requests==2.28.1
